用户导入，商品导入，导入模板 订单导出

+++ src/main/java/mall/entity/ZhongHeMallGoods.java
@@ -36,12 +36,12 @@ public class ZhongHeMallGoods {
 
     private Byte goodsSellStatus;
 
-    private Integer createUser;
+    private Long createUser;
 
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", timezone = "GMT+8")
     private Date createTime;
 
-    private Integer updateUser;
+    private Long updateUser;
 
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", timezone = "GMT+8")
     private Date updateTime;
